fix(GlobalFooter): force spinner reset and fix CheckIn spinner
For some reason the .catch cb is never invoked on Android so there's no
way for me to know if a user has pressed "cancel" or "back" instead of
selecting a location.

This solution resets the spinner state after `delay` where delay=1500 by
default. Not the best but it works well enough.