rm(RemoveCreateComponent): disabling auto-flush
I'm going to disable this until I figure out a better way. Once it
reaches a certain number of characters, it starts to lag... Perhaps
there's a bug somewhere because after leaving the page and coming back,
the lag is gone.

In any case, I've enable flushing on blur and press "done". It should be
enough to handle most cases.